chore: refactor code to different files
