+++ kadena/core/src/crypto/mod.rs
@@ -1,0 +1,7 @@
+use uint::construct_uint;
+
+pub mod hash;
+
+construct_uint! {
+    pub struct U256(4);
+}
